Adding a lot of patches from Stan Brooks.

+++ libst.h
@@ -16,9 +16,9 @@
 extern int ulaw_exp_table[256];
 extern unsigned char ulaw_comp_table[16384];
 #define st_ulaw_to_linear(ulawbyte) ulaw_exp_table[ulawbyte]
-#define st_linear_to_ulaw(linearword) ulaw_comp_table[(linearword / 4) & 0x3fff]
+#define st_linear_to_ulaw(linearword) ulaw_comp_table[((short)linearword / 4) & 0x3fff]
 #else
-unsigned char st_linear_to_ulaw( /* int sample */ );
+unsigned char st_linear_to_ulaw( /* short sample */ );
 int st_ulaw_to_linear( /* unsigned char ulawbyte */ );
 #endif
 
@@ -26,8 +26,8 @@ int st_ulaw_to_linear( /* unsigned char ulawbyte */ );
 extern int Alaw_exp_table[256];
 extern unsigned char Alaw_comp_table[16384];
 #define st_Alaw_to_linear(Alawbyte) Alaw_exp_table[Alawbyte]
-#define st_linear_to_Alaw(linearword) Alaw_comp_table[(linearword / 4) & 0x3fff]
+#define st_linear_to_Alaw(linearword) Alaw_comp_table[((short)linearword / 4) & 0x3fff]
 #else
-unsigned char st_linear_to_Alaw( /* int sample */ );
+unsigned char st_linear_to_Alaw( /* short sample */ );
 int st_Alaw_to_linear( /* unsigned char ulawbyte */ );
 #endif
